Move Filter interface out of storage/ folder
