+++ server/src/interfaces.ts
@@ -126,3 +126,11 @@ export interface MerchantPayload {
 export interface MerchantResponse extends MerchantPayload {
   id: number;
 }
+
+/**
+ * Filter Interface that contains the fields of a filter query.
+ */
+export interface Filter {
+  property: string;
+  value: any;
+}
