+++ server/src/storage/interfaces.ts
@@ -18,11 +18,3 @@
  * ResponseId type is the union type of datastore's built-in id.
  */
 export type ResponseId = number | Long | string | null;
-
-/**
- * Filter Interface that contains the fields of a filter query for datastore.
- */
-export interface Filter {
-  property: string;
-  value: any;
-}
